Support use of BUNDLE_GEMFILE env var
This is to work around an issue with Bundler 1.15.2 where it cannot run successfully when the gemfile path is relative as described here: https://github.com/bundler/bundler/pull/5815

The Bundler 1.15.2 issue has been fixed in 1.16+, but the next supported version of bundler by the official ruby buildpack is Bundler 2 (starting with #v201) which may not be feasible for an application to move to.

This issue does not affect running the app only the initial installation of gems.

One additional thing to note: Due to an oddity in bundler the bundle path needs to be relative to where the BUNDLE_GEMFILE is found. In the case of BUNDLE_GEMFILE=gemfiles/rails_3.gemfile bundler will assume the path is relative to the directory that houses the gemfile itself, so "gemfiles/". This is incorrect and is likely not where you want gems installed to. This is why this commit does a little relative_path determination before running "bundle install".