Use BUNDLE_SPECIFIC_PLATFORM on Bundler 1
This environment variable is required for installing `libv8-node`. Otherwise,
`bundle install` will fail with the error message "Could not find
libv8-node-15.14.0.1-x86_64-linux in any of the sources". This variable can't
be set using config:set, because it needs to be the same for the lifetime of a
slug, otherwise Bundler won't be able to find the right gems on disk when
running `bundle exec`.